Develop

+++ database/migrations/2020_08_30_015831_create_ficha_paciente_table.php
@@ -18,8 +18,8 @@ class CreateFichaPacienteTable extends Migration
             $table->unsignedBigInteger('IdPaciente');
             $table->string('LugarResidencia');
             $table->string('Direccion');
-            $table->decimal('Peso');
-            $table->decimal('Talla');
+            $table->Integer('Peso');
+            $table->Integer('Talla');
             $table->Integer('SPO2');
             $table->Integer('FC');
             $table->Integer('FR');
